requirements.txt
streamlit
PyPDF2
transformers
sentence-transformers
pytesseract
pdf2image
huggingface_hub
numpy
python-dotenv
